requirements.txt
Flask==2.3.2
line-bot-sdk==2.2.1
